+++ sticky_mitten_avatar/sma_controller.py
@@ -209,8 +209,6 @@ class StickyMittenAvatarController(FloorplanController):
                      "enabled": False},
                     {"$type": "set_shadow_strength",
                      "strength": 1.0},
-                    {"$type": "set_sleep_threshold",
-                     "sleep_threshold": 0.01},
                     {"$type": "set_screen_size",
                      "width": screen_width,
                      "height": screen_height},
@@ -337,17 +335,18 @@ class StickyMittenAvatarController(FloorplanController):
 
         self._end_task()
 
-    def _end_task(self, toggle_sensor: bool = True) -> None:
+    def _end_task(self, enable_sensor: bool = True) -> None:
         """
         End the task and update the frame data.
 
-        :param toggle_sensor: If True, toggle the image sensor.
+        :param enable_sensor: If True, enable the image sensor.
         """
 
         commands = [self._avatar.get_default_sticky_mitten_profile()]
 
-        if not self._demo and toggle_sensor:
-            commands.append({"$type": "toggle_image_sensor",
+        if enable_sensor:
+            commands.append({"$type": "enable_image_sensor",
+                             "enable": True,
                              "sensor_name": "SensorContainer",
                              "avatar_id": self._avatar.id})
         # Request output data to update the frame data.
@@ -427,7 +426,7 @@ class StickyMittenAvatarController(FloorplanController):
 
     def reach_for_target(self, arm: Arm, target: Dict[str, float], do_motion: bool = True,
                          check_if_possible: bool = True, stop_on_mitten_collision: bool = True,
-                         precision: float = 0.05) -> TaskStatus:
+                         precision: float = 0.05, absolute: bool = False) -> TaskStatus:
         """
         Bend an arm joints of an avatar to reach for a target position.
 
@@ -441,11 +440,12 @@ class StickyMittenAvatarController(FloorplanController):
         - `mitten_collision` (If `stop_if_mitten_collision == True`)
 
         :param arm: The arm (left or right).
-        :param target: The target position for the mitten relative to the avatar.
+        :param target: The target position for the mitten.
         :param do_motion: If True, advance simulation frames until the pick-up motion is done.
         :param stop_on_mitten_collision: If true, the arm will stop bending if the mitten collides with an object other than the target object.
         :param check_if_possible: If True, before bending the arm, check if the mitten can reach the target assuming no obstructions; if not, don't try to bend the arm.
         :param precision: The precision of the action. If the mitten is this distance or less away from the target position, the action returns `success`.
+        :param absolute: If True, `target` is in absolute world coordinates. If False, `target` is in coordinates relative to the avatar's position and rotation.
 
         :return: A `TaskStatus` indicating whether the avatar can reach the target and if not, why.
         """
@@ -453,6 +453,10 @@ class StickyMittenAvatarController(FloorplanController):
         self._start_task()
 
         target = TDWUtils.vector3_to_array(target)
+
+        # Convert to relative coordinates.
+        if absolute:
+            target = self._avatar.get_rotated_target(target=target)
 
         # Check if it is possible for the avatar to reach the target.
         if check_if_possible:
@@ -601,20 +605,23 @@ class StickyMittenAvatarController(FloorplanController):
             if not done:
                 self.communicate([])
 
-    def _stop_avatar(self) -> None:
+    def _stop_avatar(self, enable_sensor: bool) -> None:
         """
         Advance 1 frame and stop the avatar's movement and turning.
+
+        :param enable_sensor: If True, enable the image sensor.
         """
 
         self.communicate({"$type": "set_avatar_drag",
                           "drag": self._STOP_DRAG,
                           "angular_drag": self._STOP_DRAG,
                           "avatar_id": self._avatar.id})
-        self._end_task()
+        self._end_task(enable_sensor=enable_sensor)
         self._avatar.status = TaskStatus.idle
 
     def turn_to(self, target: Union[Dict[str, float], int], force: float = 1000,
-                stopping_threshold: float = 0.15, num_attempts: int = 200) -> TaskStatus:
+                stopping_threshold: float = 0.15, num_attempts: int = 200,
+                enable_sensor_on_finish: bool = True) -> TaskStatus:
         """
         Turn the avatar to face a target position or object.
 
@@ -627,6 +634,7 @@ class StickyMittenAvatarController(FloorplanController):
         :param force: The force at which the avatar will turn. More force = faster, but might overshoot the target.
         :param stopping_threshold: Stop when the avatar is within this many degrees of the target.
         :param num_attempts: The avatar will apply more angular force this many times to complete the turn before giving up.
+        :param enable_sensor_on_finish: Enable the camera upon completing the task. This is for internal use only.
 
         :return: A `TaskStatus` indicating whether the avatar turned successfully and if not, why.
         """
@@ -666,20 +674,16 @@ class StickyMittenAvatarController(FloorplanController):
 
         self._avatar.status = TaskStatus.ongoing
 
-        # Set a low drag.
-        self.communicate({"$type": "set_avatar_drag",
-                          "drag": 0,
-                          "angular_drag": 0.05,
-                          "avatar_id": self._avatar.id})
-
-        turn_command = {"$type": "turn_avatar_by",
-                        "torque": force * direction,
-                        "avatar_id": self._avatar.id}
-
-        # Begin to turn.
-        self.communicate(turn_command)
         i = 0
         while i < num_attempts:
+            self.communicate([{"$type": "set_avatar_drag",
+                               "drag": 0,
+                               "angular_drag": 0.05,
+                               "avatar_id": self._avatar.id},
+                              self._avatar.get_rotation_sticky_mitten_profile(),
+                              {"$type": "turn_avatar_by",
+                               "torque": force * direction,
+                               "avatar_id": self._avatar.id}])
             # Coast to a stop.
             coasting = True
             while coasting:
@@ -687,27 +691,26 @@ class StickyMittenAvatarController(FloorplanController):
                 state, previous_angle = _get_turn_state()
                 # The turn succeeded!
                 if state == TaskStatus.success:
-                    self._stop_avatar()
+                    self._stop_avatar(enable_sensor_on_finish)
                     return state
                 # The turn failed.
                 elif state != TaskStatus.ongoing:
-                    self._stop_avatar()
+                    self._stop_avatar(enable_sensor_on_finish)
                     return state
                 self.communicate([])
 
             # Turn.
-            self.communicate(turn_command)
             state, previous_angle = _get_turn_state()
             # The turn succeeded!
             if state == TaskStatus.success:
-                self._stop_avatar()
+                self._stop_avatar(enable_sensor_on_finish)
                 return state
             # The turn failed.
             elif state != TaskStatus.ongoing:
-                self._stop_avatar()
+                self._stop_avatar(enable_sensor_on_finish)
                 return state
             i += 1
-        self._stop_avatar()
+        self._stop_avatar(enable_sensor_on_finish)
         return TaskStatus.too_long
 
     def turn_by(self, angle: float, force: float = 1000, stopping_threshold: float = 0.15, num_attempts: int = 200) -> \
@@ -807,43 +810,43 @@ class StickyMittenAvatarController(FloorplanController):
         if turn:
             # Turn to the target.
             status = self.turn_to(target=TDWUtils.array_to_vector3(target), force=turn_force,
-                                  stopping_threshold=turn_stopping_threshold, num_attempts=num_attempts)
+                                  stopping_threshold=turn_stopping_threshold, num_attempts=num_attempts,
+                                  enable_sensor_on_finish=False)
             if status != TaskStatus.success:
-                self._stop_avatar()
+                self._stop_avatar(True)
                 return status
         self._start_task()
         self._avatar.status = TaskStatus.ongoing
-
-        # Go to the target.
-        self.communicate({"$type": "set_avatar_drag",
-                          "drag": 0.1,
-                          "angular_drag": 100,
-                          "avatar_id": self._avatar.id})
         i = 0
         while i < num_attempts:
             # Start gliding.
-            self.communicate({"$type": "move_avatar_forward_by",
-                              "magnitude": move_force,
-                              "avatar_id": self._avatar.id})
+            self.communicate([{"$type": "move_avatar_forward_by",
+                               "magnitude": move_force,
+                               "avatar_id": self._avatar.id},
+                              {"$type": "set_avatar_drag",
+                               "drag": 0.1,
+                               "angular_drag": 100,
+                               "avatar_id": self._avatar.id},
+                              self._avatar.get_movement_sticky_mitten_profile()])
             t = _get_state()
             if t == TaskStatus.success:
-                self._stop_avatar()
+                self._stop_avatar(True)
                 return t
             elif t != TaskStatus.ongoing:
-                self._stop_avatar()
+                self._stop_avatar(True)
                 return t
             # Glide.
             while np.linalg.norm(self._avatar.frame.get_velocity()) > 0.1:
                 self.communicate([])
                 t = _get_state()
                 if t == TaskStatus.success:
-                    self._stop_avatar()
+                    self._stop_avatar(True)
                     return t
                 elif t != TaskStatus.ongoing:
-                    self._stop_avatar()
+                    self._stop_avatar(True)
                     return t
             i += 1
-        self._stop_avatar()
+        self._stop_avatar(True)
         return TaskStatus.too_long
 
     def move_forward_by(self, distance: float, move_force: float = 80, move_stopping_threshold: float = 0.35,
@@ -945,14 +948,17 @@ class StickyMittenAvatarController(FloorplanController):
         self._end_task()
         return TaskStatus.success
 
-    def put_in_container(self, object_id: int, container_id: int, arm: Arm, num_attempts: int = 10) -> TaskStatus:
+    def put_in_container(self, object_id: int, container_id: int, arm: Arm) -> TaskStatus:
         """
         Try to put an object in a container.
 
         1. The avatar will grasp the object and a container via `grasp_object()` if it isn't holding them already.
-        2. The avatar will lift the object up and then over the container via `reach_for_target()`
-        3. The avatar will make multiple attempts to position the object over the container via `reach_for_target()` plus some backend-only logic.
-        4. The avatar will `drop()` the object into the container.
+        2. The avatar will lift the object up.
+        3. The container and its contents will be teleported to be in front of the avatar.
+        4. The avatar will move the object over the container and drop it.
+        5. The avatar will pick up the container again.
+
+        The container will be teleport to
 
         Possible [return values](task_status.md):
 
@@ -971,7 +977,6 @@ class StickyMittenAvatarController(FloorplanController):
         :param object_id: The ID of the object that the avatar will try to put in the container.
         :param container_id: The ID of the container. To determine if an object is a container, see [`StaticObjectInfo.container')(static_object_info.md).
         :param arm: The arm that will try to pick up the object.
-        :param num_attempts: Make this many attempts to re-position the object above the container.
 
         :return: A `TaskStatus` indicating whether the avatar put the object in the container and if not, why.
         """
@@ -979,11 +984,11 @@ class StickyMittenAvatarController(FloorplanController):
         if not self.static_object_info[container_id].container:
             return TaskStatus.not_a_container
 
-        self._start_task()
+        self._stop_avatar(enable_sensor=False)
 
         # A "full" container has too many objects such that physics might glitch.
         overlap_ids = self._get_objects_in_container(container_id=container_id)
-        if len(overlap_ids) > 4:
+        if len(overlap_ids) > 3:
             self._end_task()
             return TaskStatus.full_container
 
@@ -991,62 +996,79 @@ class StickyMittenAvatarController(FloorplanController):
         if object_id not in self.frame.held_objects[arm]:
             status = self.grasp_object(object_id=object_id, arm=arm)
             if status != TaskStatus.success:
-                self._end_task()
+                self._end_task(enable_sensor=False)
                 return status
         container_arm = Arm.left if arm == Arm.right else Arm.right
 
-        # These values will be used to initially lift the object and then to re-aim it until it's over the container.
-        d_arm_z = -0.1
-        target_y = 0.6
-
         # Lift up the object.
-        self.reach_for_target(target={"x": 0.35 if arm == Arm.right else -0.35, "y": target_y, "z": 0.36},
+        self.reach_for_target(target={"x": 0.35 if arm == Arm.right else -0.35, "y": 0.3, "z": 0.36},
                               arm=arm,
                               check_if_possible=False,
                               stop_on_mitten_collision=False,
                               precision=0.2)
-        # Lift the container.
-        self.reach_for_target(target={"x": 0.05 if arm == Arm.right else -0.05, "y": 0.05, "z": 0.32},
-                              arm=container_arm,
+
+        # Let the container fall to the ground.
+        self.drop(arm=container_arm)
+        self.reset_arm(arm=container_arm)
+
+        # Try to nudge the container to be directly in front of the avatar.
+        new_container_position = self.frame.avatar_transform.position + np.array([-0.215 if arm == Arm.right else 0.215,
+                                                                                  0, 0.341])
+        new_container_angle = get_angle(forward=self.frame.avatar_transform.forward,
+                                        origin=self.frame.avatar_transform.position,
+                                        position=new_container_position)
+        new_container_position = rotate_point_around(point=new_container_position,
+                                                     origin=self.frame.avatar_transform.position,
+                                                     angle=new_container_angle)
+        new_container_rotation = TDWUtils.array_to_vector4(self.frame.avatar_transform.rotation)
+        # Check if there's anything in the way.
+        resp = self.communicate({"$type": "send_overlap_box",
+                                 "position": TDWUtils.array_to_vector3(new_container_position),
+                                 "rotation": new_container_rotation,
+                                 "half_extents": TDWUtils.array_to_vector3(self.static_object_info[container_id].size /
+                                                                           2)})
+        overlap_ids = get_data(resp=resp, d_type=Overlap).get_object_ids()
+        can_nudge = True
+        for overlap_id in overlap_ids:
+            if not self.static_object_info[overlap_id].container and \
+                    not self.static_object_info[overlap_id].target_object:
+                can_nudge = False
+                break
+        # Teleport the objects an easily-to-reach location.
+        if can_nudge:
+            self._end_task(enable_sensor=False)
+            overlap_ids = self._get_objects_in_container(container_id=container_id)
+            delta_position = new_container_position - self.frame.object_transforms[container_id].position
+            overlap_ids.append(container_id)
+            teleport_commands = [{"$type": "rotate_object_to",
+                                  "rotation": TDWUtils.array_to_vector4(self.frame.avatar_transform.rotation),
+                                  "id": container_id,
+                                  "physics": True}]
+            for teleport_id in overlap_ids:
+                teleport_position = TDWUtils.array_to_vector3(self.frame.object_transforms[teleport_id].position +
+                                                              delta_position)
+                teleport_position["y"] += 0.03
+                teleport_commands.append({"$type": "teleport_object",
+                                          "position": teleport_position,
+                                          "id": teleport_id,
+                                          "physics": True})
+            self.communicate(teleport_commands)
+            self._wait_for_objects_to_stop(object_ids=overlap_ids)
+            self._end_task()
+
+        # Lift the arm away.
+        self.reach_for_target(target={"x": 0.25 if arm == Arm.right else -0.25, "y": 0.6, "z": 0.3},
+                              arm=arm,
                               check_if_possible=False,
                               stop_on_mitten_collision=False)
-        # Twist the wrist of the arm holding the container.
-        self._roll_wrist(arm=container_arm, angle=60)
+        aim_position = self.frame.object_transforms[container_id].position
+        aim_position[1] = 0.3
+        self.reach_for_target(arm=arm,
+                              target={"x": -0.091 if arm == Arm.right else 0.091, "y": 0.306, "z": 0.392},
+                              stop_on_mitten_collision=False,
+                              check_if_possible=False)
 
-        self._end_task(toggle_sensor=False)
-        container_position = self.frame.object_transforms[container_id].position
-
-        # Continuously try to position the object over the container.
-        hit = False
-        attempts = 0
-        target_position = self._avatar.get_rotated_target(target=container_position)
-        target_x = target_position[0]
-        target_z = target_position[2]
-        while not hit and attempts < num_attempts:
-            # Call `_end_task()` to update the FrameData.
-            self._end_task(toggle_sensor=False)
-
-            # Get the new distance.
-            container_position = self.frame.object_transforms[container_id].position
-
-            # Raycast from the object directly downward to see if it's directly above the container.
-            object_position = TDWUtils.array_to_vector3(self.frame.object_transforms[object_id].position)
-
-            resp = self.communicate({"$type": "send_raycast",
-                                     "origin": object_position,
-                                     "destination": {"x": object_position["x"], "y": 0, "z": object_position["z"]}})
-            raycast = get_data(resp=resp, d_type=Raycast)
-            hit = raycast.get_hit_object()
-            hit_point = raycast.get_point()
-            distance_to_container = np.linalg.norm(container_position - hit_point)
-            # Move the arm slightly until there's a hit that is close to the center of the container.
-            if not hit or distance_to_container > 0.2:
-                target_z += d_arm_z
-                self.reach_for_target(target={"x": target_x, "y": target_y, "z": target_z}, arm=arm,
-                                      precision=0.2, check_if_possible=False)
-                # Roll the wrist of the arm holding the object so that the mitten isn't in the way.
-                self._roll_wrist(arm=arm, angle=0, precision=0.05)
-            attempts += 1
+        self._end_task(enable_sensor=False)
 
         # Drop the object.
         self.drop(arm=arm, reset_arm=False, do_motion=False)
@@ -1057,26 +1079,39 @@ class StickyMittenAvatarController(FloorplanController):
                               check_if_possible=False,
                               stop_on_mitten_collision=False)
 
-        # Wait for the object to stop moving.
-        self.communicate({"$type": "send_rigidbodies",
-                          "frequency": "always"})
-        sleeping = False
-        while not sleeping:
-            # Advance one frame.
-            resp = self.communicate([])
-            rigidbodies = get_data(resp=resp, d_type=Rigidbodies)
-            # Check if the object stopped moving.
-            for i in range(rigidbodies.get_num()):
-                if rigidbodies.get_id(i) == object_id:
-                    sleeping = rigidbodies.get_sleeping(i) or np.linalg.norm(rigidbodies.get_velocity(i)) < 0.1
-                    break
+        self._wait_for_objects_to_stop(object_ids=[object_id])
+        self.reset_arm(arm=arm)
+
+        # Check if the object is in the container.
         overlap_ids = self._get_objects_in_container(container_id=container_id)
         if object_id not in overlap_ids:
-            self._end_task()
             return TaskStatus.not_in_container
-        else:
-            self._end_task()
-            return TaskStatus.success
+
+        self.reset_arm(arm=container_arm)
+
+        # Move the container and its objects in front of the mitten.
+        mitten_id = self._avatar.mitten_ids[container_arm]
+        new_container_position = self.frame.avatar_body_part_transforms[mitten_id].position + self.frame.\
+            avatar_transform.forward * 0.3
+        delta_position = new_container_position - self.frame.object_transforms[container_id].position
+        teleport_commands = [{"$type": "teleport_object",
+                              "id": container_id,
+                              "position": TDWUtils.array_to_vector3(new_container_position),
+                              "physics": True}]
+        for overlap_id in overlap_ids:
+            teleport_position = self.frame.object_transforms[overlap_id].position + delta_position
+            teleport_position[1] += 0.03
+            teleport_commands.append({"$type": "teleport_object",
+                                      "id": overlap_id,
+                                      "position": TDWUtils.array_to_vector3(teleport_position),
+                                      "physics": True})
+
+        self.communicate(teleport_commands)
+
+        self._wait_for_objects_to_stop(object_ids=[object_id])
+
+        # Pick up the container again.
+        return self.grasp_object(object_id=container_id, arm=container_arm, check_if_possible=False)
 
     def pour_out_container(self, arm: Arm) -> TaskStatus:
         """
@@ -1096,26 +1131,6 @@ class StickyMittenAvatarController(FloorplanController):
 
         :return: A `TaskStatus` indicating whether the avatar poured all objects out of the container and if not, why.
         """
-
-        def _wait_until_objects_stop() -> None:
-            """
-            Wait until the objects in the container stop moving.
-            """
-
-            # Wait for the objects to stop moving.
-            resp = self.communicate({"$type": "send_rigidbodies",
-                                     "frequency": "always",
-                                     "ids": overlap_ids})
-            moving = True
-            while moving:
-                rigidbodies = get_data(resp=resp, d_type=Rigidbodies)
-                moving = False
-                for i in range(rigidbodies.get_num()):
-                    sleeping = rigidbodies.get_sleeping(i) or np.linalg.norm(rigidbodies.get_velocity(i)) < 0.1
-                    if not sleeping:
-                        moving = True
-                        break
-                resp = self.communicate([])
 
         # Make sure that this arm is holding a container.
         held = self._avatar.frame.get_held_left() if arm == Arm.left else self._avatar.frame.get_held_right()
@@ -1138,10 +1153,10 @@ class StickyMittenAvatarController(FloorplanController):
         self._roll_wrist(arm=arm, angle=90)
         # Lift the arm to tilt the container.
         self.reach_for_target(arm=arm, target={"x": -0.3 if arm == Arm.left else 0.3, "y": 0.6, "z": 0.35},
-                              check_if_possible=False, stop_on_mitten_collision=False, precision=0.2)
+                              check_if_possible=False, stop_on_mitten_collision=False)
         self._roll_wrist(arm=arm, angle=90)
 
-        _wait_until_objects_stop()
+        self._wait_for_objects_to_stop(object_ids=overlap_ids)
 
         # Get all of the objects in the container. If there aren't any, this task succeeded.
         overlap_ids = self._get_objects_in_container(container_id=container_id)
@@ -1151,7 +1166,7 @@ class StickyMittenAvatarController(FloorplanController):
         # Try to shake objects out of the container.
         else:
             self.shake(joint_name=f"elbow_{arm.name}", num_shakes=(2, 2))
-            _wait_until_objects_stop()
+            self._wait_for_objects_to_stop(object_ids=overlap_ids)
             overlap_ids = self._get_objects_in_container(container_id=container_id)
             self._end_task()
             return TaskStatus.success if len(overlap_ids) == 0 else TaskStatus.still_in_container
@@ -1230,7 +1245,8 @@ class StickyMittenAvatarController(FloorplanController):
                              "avatar_id": cam_id})
         if images == "cam":
             # Disable avatar cameras.
-            commands.append({"$type": "toggle_image_sensor",
+            commands.append({"$type": "enable_image_sensor",
+                             "enable": False,
                              "sensor_name": "SensorContainer"})
 
             commands.append({"$type": "send_images",
@@ -1275,12 +1291,12 @@ class StickyMittenAvatarController(FloorplanController):
         # Decide which overlap shape to use depending on the container shape.
         if shape == "box":
             resp = self.communicate({"$type": "send_overlap_box",
-                                     "position": center,
+                                     "position": pos,
                                      "rotation": TDWUtils.array_to_vector4(rot),
-                                     "half_extents": TDWUtils.array_to_vector3(size / 2)})
+                                     "half_extents": TDWUtils.array_to_vector3(size * 1.25)})
         elif shape == "sphere":
             resp = self.communicate({"$type": "send_overlap_sphere",
-                                     "position": center,
+                                     "position": pos,
                                      "radius": min(size)})
         elif shape == "capsule":
             resp = self.communicate({"$type": "send_overlap_capsule",
@@ -1538,13 +1554,15 @@ class StickyMittenAvatarController(FloorplanController):
                           "angular_drag": self._STOP_DRAG},
                          {"$type": "set_pass_masks",
                           "pass_masks": pass_masks},
-                         {"$type": "toggle_image_sensor",
+                         {"$type": "enable_image_sensor",
+                          "enable": False,
                           "sensor_name": "FollowCamera"},
                          {"$type": "teleport_avatar_to",
                           "avatar_id": "a",
                           "position": avatar_position}])
         if not self._demo:
-            commands.append({"$type": "toggle_image_sensor",
+            commands.append({"$type": "enable_image_sensor",
+                             "enable": False,
                              "sensor_name": "SensorContainer"})
         # Set all sides of both mittens to be sticky.
         for sub_mitten in ["palm", "back", "side"]:
@@ -1593,6 +1611,33 @@ class StickyMittenAvatarController(FloorplanController):
         if self._demo or self._avatar is None:
             return
 
-        self._avatar_commands.append({"$type": "toggle_image_sensor",
+        self._avatar_commands.append({"$type": "enable_image_sensor",
+                                      "enable": False,
                                       "sensor_name": "SensorContainer",
                                       "avatar_id": self._avatar.id})
+
+    def _wait_for_objects_to_stop(self, object_ids: List[int]) -> None:
+        """
+        Wait for some objects to stop moving.
+
+        :param object_ids: A list of object IDs.
+        """
+
+        # Request rigidbody data per frame for each of the objects.
+        resp = self.communicate({"$type": "send_rigidbodies",
+                                 "frequency": "always",
+                                 "ids": object_ids})
+        sleeping = False
+        while not sleeping:
+            sleeping = True
+            # Advance one frame.
+            rigidbodies = get_data(resp=resp, d_type=Rigidbodies)
+            # Check if the object stopped moving.
+            for i in range(rigidbodies.get_num()):
+                # Check if this object is moving.
+                if np.linalg.norm(rigidbodies.get_velocity(i)) > 0.1:
+                    sleeping = False
+                    break
+            resp = self.communicate([])
+
+        self._end_task(enable_sensor=False)